adding full crud functionality

+++ db/migrate/20160605025113_create_athletes.rb
@@ -2,8 +2,8 @@ class CreateAthletes < ActiveRecord::Migration
   def change
     create_table :athletes do |t|
       t.string :name
-      t.float :height
-      t.float :weight
+      t.integer :height
+      t.integer :weight
 
       t.timestamps null: false
     end
